requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.16.1
numpy>=1.26.0
Pillow>=10.1.0
pydantic==2.5.0
streamlit>=1.28.0
matplotlib>=3.7.0
pandas>=2.0.0
opencv-python-headless>=4.8.0



